+++ docs/make.jl
@@ -1,11 +1,12 @@
 
-using Base: Docs.Text, get_extension
+using Base: get_extension
 using Dates: Dates
 using Documenter: HTML, DocMeta.setdocmeta!, deploydocs, makedocs
-using LaTeXStrings: LaTeXStrings
+using LaTeXStrings: LaTeXStrings, @L_str
 using Luxor: Drawing, finish, julia_blue, julia_green, julia_purple, julia_red, rect, sethue
+using Markdown: Markdown, @md_str
 using Typstry
-using Typstry: Typst, enclose, examples, join_with, preamble, typst_mime
+using Typstry: Typst, _show_typst, enclose, join_with, preamble, typst_mime
 
 const assets = joinpath(@__DIR__, "src", "assets")
 const _examples = joinpath(assets, "examples.typ")
@@ -13,6 +14,35 @@ const logo = joinpath(assets, "logo.svg")
 const modes = instances(Mode)
 const width, height = 210, 297
 const modules = [Typstry]
+const examples = [
+    Dates.Date(1) => Dates.Date,
+    Dates.DateTime(1) => Dates.DateTime,
+    Dates.Time(0) => Dates.Time,
+    html"<p>a</p>" => Docs.HTML,
+    text"[\"a\"]" => Docs.Text,
+    L"a" => LaTeXStrings.LaTeXString,
+    md"# A" => Markdown.MD,
+    [true, 1, Any[1.2, 1 // 2]] => AbstractArray,
+    'a' => AbstractChar,
+    1.2 => AbstractFloat,
+    Any[true 1; 1.2 1 // 2] => AbstractMatrix,
+    "a" => AbstractString,
+    true => Bool,
+    im => Complex{Bool},
+    1 + 2im => Complex,
+    π => Irrational,
+    nothing => Nothing,
+    0:2:6 => OrdinalRange{<:Integer, <:Integer},
+    1 // 2 => Rational,
+    r"[a-z]" => Regex,
+    1 => Signed,
+    StepRangeLen(0, 2, 4) => StepRangeLen{<:Integer, <:Integer, <:Integer},
+    (true, 1, 1.2, 1 // 2) => Tuple,
+    Typst(1) => Typst,
+    typst"[\"a\"]" => TypstString,
+    TypstText([1, 2, 3, 4]) => TypstText,
+    0xff => Unsigned
+]
 
 mkpath(assets)
 Drawing(width, height, :svg, logo)
@@ -49,24 +79,37 @@ open(_examples; truncate = true) do file
     join(file, map(mode -> "`$mode`", modes), ", ")
     println(file, "\n    ),")
 
-    join_with(file, examples, ",\n") do file, (v, t)
+    join_with(file, examples, ",\n") do file, example
+        v, t = example
         is_vector = v isa Vector
 
         print(file, "    ")
+        show(file,
+            if is_vector "[true, 1, Any[1.2, 1//2]]"
+            elseif v isa Dates.Date "Dates.Date(1)"
+            elseif v isa Dates.DateTime "Dates.DateTime(1)"
+            elseif v isa Dates.Time "Dates.Time(1)"
+            elseif v isa Docs.HTML "html\"<p>a</p>\""
+            elseif v isa Docs.Text "text\"[\\\"a\\\"]\""
+            elseif v isa LaTeXStrings.LaTeXString "L\"a\""
+            elseif v isa Markdown.MD "md\"# a\""
+            elseif v isa StepRangeLen "StepRangeLen(0, 2, 4)"
+            elseif v isa Typst "Typst(1)"
+            elseif v isa TypstText "TypstText([1, 2, 3, 4])"
+            else repr(v)
+            end
+        )
+        print(file, ", `")
 
-        if is_vector print(file, "\"[true, 1, Any[1.2, 1//2]]\"")
-        elseif v isa StepRangeLen print(file, "\"StepRangeLen(0, 2, 4)\"")
-        elseif v isa Text print(file, "\"text\\\"[\\\\\\\"a\\\\\\\"]\\\"\"")
-        elseif v isa Typst print(file, "\"Typst(1)\"")
-        elseif v isa TypstText print(file, "\"TypstText([1, 2, 3, 4])\"")
-        else show(file, repr(v))
+        if v isa Docs.HTML print(file, "Docs.HTML")
+        elseif v isa Docs.Text print(file, "Docs.Text")
+        else print(file, t)
         end
 
-        print(file, ", `", v isa Text ? "Docs.Text" : t, "`,", is_vector || v isa AbstractMatrix ? "\n        " : " ")
+        print(file, "`,", is_vector || v isa AbstractMatrix ? "\n        " : " ")
         join_with(file, modes, ", ") do file, mode
-            enclose((file, v) ->
-                show(IOContext(file, :mode => mode, :depth => 2), typst_mime, Typst(v)),
-            file, v, (mode == math ? ("\$", "\$") : ("[" * (mode == code ? "#" : ""), "]"))...)
+            enclose((file, v) -> _show_typst(IOContext(file, :mode => mode, :depth => 2), v),
+                file, v, (mode == math ? ("\$", "\$") : ("[" * (mode == code ? "#" : ""), "]"))...)
         end
     end
 
@@ -83,20 +126,15 @@ for extension in [:Dates, :LaTeXStrings, :Markdown]
     push!(modules, _module)
 end
 
-makedocs(;
-    modules,
-    sitename = "Typstry.jl",
-    format = HTML(edit_link = "main"),
-    pages = [
-        "Home" => "index.md",
-        "Getting Started" => "getting_started.md",
-        "Tutorials" => ["Interface" => "tutorials/interface.md"],
-        "Manual" => map(page -> uppercasefirst(page) => joinpath("manual", page * ".md"),
-            ["strings", "commands", "extensions", "internals"])
-    ]
-)
+makedocs(; modules, sitename = "Typstry.jl", format = HTML(edit_link = "main"), pages = [
+    "Home" => "index.md",
+    "Getting Started" => "getting_started.md",
+    "Tutorials" => ["Interface" => "tutorials/interface.md"],
+    "Manual" => map(page -> uppercasefirst(page) => joinpath("manual", page * ".md"),
+        ["strings", "commands", "extensions", "internals"])
+])
 
-deploydocs(
+deploydocs(;
     devbranch = "main",
     devurl = "development",
     repo = "github.com/jakobjpeters/Typstry.jl.git"