+++ ext/DatesExtension.jl
@@ -2,7 +2,7 @@
 module DatesExtension
 
 import Typstry: show_typst
-using Dates: Date, Time, DateTime
+using Dates: Date, Time, DateTime, day, hour, minute, month, second, year
 using Typstry: _show_typst, code_mode, enclose, join_with
 
 # Internals
@@ -17,10 +17,23 @@ function show_date_time(io, x, fs...)
     end, io, x, "datetime(", ")"; fs)
 end
 
+show_dates(io, x::Date) = show_date_time(io, x, year, month, day)
+show_dates(io, x::Time) = show_date_time(io, x, hour, minute, second)
+show_dates(io, x::DateTime) = show_date_time(io, x, year, month, day, hour, minute, second)
+
 # Strings
 
-show_typst(io, x::Date) = show_date_time(io, x, year, month, day)
-show_typst(io, x::Time) = show_date_time(io, x, hour, minute, second)
-show_typst(io, x::DateTime) = show_date_time(io, x, year, month, day, hour, minute, second)
+"""
+    show_typst(io, ::Union{Date, Time, DateTime})
+
+Print in Typst format for Dates.jl.
+
+| Type       | Settings | Parameters |
+|:-----------|:---------|:-----------|
+| `Date`     | `:mode`  |            |
+| `Time`     | `:mode`  |            |
+| `DateTime` | `:mode`  |            |
+"""
+show_typst(io, x::Union{Date, Time, DateTime}) = show_dates(io, x)
 
 end # module