+++ src/Typstry.jl
@@ -2,11 +2,11 @@
 module Typstry
 
 import Base:
-    Docs.HTML, IOBuffer, ==, addenv, codeunit, detach, eltype, firstindex,
+    IOBuffer, ==, addenv, codeunit, detach, eltype, firstindex,
     getindex, hash, ignorestatus, isvalid, iterate, keys, lastindex,
     length, ncodeunits, pointer, repr, run, setenv, show, showerror
 using Artifacts: @artifact_str
-using Base: Docs.Text, Iterators.Stateful, Meta.parse, escape_raw_string
+using Base: Docs.HTML, Docs.Text, Iterators.Stateful, Meta.parse, escape_raw_string
 using PrecompileTools: PrecompileTools, @compile_workload
 using Typst_jll: typst