+++ HISTORY.md
@@ -49,7 +49,7 @@
 - `show_typst` prints a value in Typst format, with Julia settings and Typst parameters given in an `IOContext`
     - Implemented for `AbstractChar`, `AbstractFloat`, `AbstractMatrix`, `AbstractString`, `AbstractVector`, `Bool`, `Complex`, `Irrational`, `Nothing`, `OrdinalRange{<:Integer, <:Integer}`, `Rational`, `Regex`, `Signed`, `StepRangeLen{<:Integer, <:Integer, <:Integer}`, `Text`, and `TypstString`
 - `show(::IO, ::MIME"text/typst", ::Any)` provides default settings to `show_typst`
-- `Mode` is a setting that specifies the current Typst context
+- `Mode` is used to specify the current Typst context in `show_typst`
 - `TypstString` is an `AbstractString` implementing the `String` interface and uses `show(::IO, ::MIME"text/typst", ::Any)` to print values to Typst format
 - `@typst_str` constructs `TypstString`s and supports formatted interpolation
 - `typst_text` constructs a `TypstString` using `print` instead of `show(::IO, ::MIME"text/typst", ::Any)`